requirements.txt
Django==1.11.9
djangorestframework==3.7.7
pytz==2017.3
gunicorn